Make the "Buy a Ticket" button font consistent across platforms.
The "Buy a Ticket" button currently uses a font stack based on Rockwell. It
is a lovely font, but it's not available everywhere. The button looks quite
different on Android, desktop Linux, and anecdotally on Macs. I'm sure it
looks different on Windows too.

This pull request changes the font stack to fall back to Rokkitt, from the
Google Web Fonts collection. As best I can tell, it's quite a close subsitute
for Rockwell.

Compare:
  http://www.fonts.com/font/monotype/rockwell
  https://www.google.com/fonts/specimen/Rokkitt

Closes #45